remove color configuration & caching. All colors are calculated every time
