+++ src/config.rs
@@ -39,13 +39,6 @@ impl Settings {
     }
 }
 
-// macro_rules! update_attribute {
-//     ($self_attr:expr, $yaml:ident, $key:expr) => {
-//         if let Some(attr) = read_yaml_value($yaml, $key) {
-//             $self_attr = attr;
-//         }
-//     };
-// }
 /// Holds every configurable aspect of the application.
 /// All attributes are hardcoded then updated from optional values
 /// of the config file.
@@ -95,86 +88,6 @@ impl Config {
         &self.terminal
     }
 }
-
-// fn read_yaml_value(yaml: &serde_yaml::value::Value, key: &str) -> Option<String> {
-//     yaml[key].as_str().map(|s| s.to_string())
-// }
-
-// /// Holds configurable colors for every kind of file.
-// /// "Normal" files are displayed with a different color by extension.
-// #[derive(Debug, Clone)]
-// pub struct Colors {
-//     /// Color for `directory` files.
-//     pub directory: String,
-//     /// Color for `block` files.
-//     pub block: String,
-//     /// Color for `char` files.
-//     pub char: String,
-//     /// Color for `fifo` files.
-//     pub fifo: String,
-//     /// Color for `socket` files.
-//     pub socket: String,
-//     /// Color for `symlink` files.
-//     pub symlink: String,
-//     /// Color for broken `symlink` files.
-//     pub broken: String,
-//     // /// Colors for normal files, depending of extension
-//     // pub color_cache: ColorCache,
-// }
-//
-// impl Colors {
-//     fn update_from_config(&mut self, yaml: &serde_yaml::value::Value) {
-//         update_attribute!(self.directory, yaml, "directory");
-//         update_attribute!(self.block, yaml, "block");
-//         update_attribute!(self.char, yaml, "char");
-//         update_attribute!(self.fifo, yaml, "fifo");
-//         update_attribute!(self.socket, yaml, "socket");
-//         update_attribute!(self.symlink, yaml, "symlink");
-//         update_attribute!(self.broken, yaml, "broken");
-//     }
-//
-//     fn new() -> Self {
-//         Self {
-//             directory: "red".to_owned(),
-//             block: "yellow".to_owned(),
-//             char: "green".to_owned(),
-//             fifo: "blue".to_owned(),
-//             socket: "cyan".to_owned(),
-//             symlink: "magenta".to_owned(),
-//             broken: "white".to_owned(),
-//             // color_cache: ColorCache::default(),
-//         }
-//     }
-// }
-//
-// impl Default for Colors {
-//     fn default() -> Self {
-//         Self::new()
-//     }
-// }
-//
-// /// Convert a string color into a `tuikit::Color` instance.
-// pub fn str_to_tuikit(color: &str) -> Color {
-//     match color {
-//         "white" => Color::WHITE,
-//         "red" => Color::RED,
-//         "green" => Color::GREEN,
-//         "blue" => Color::BLUE,
-//         "yellow" => Color::YELLOW,
-//         "cyan" => Color::CYAN,
-//         "magenta" => Color::MAGENTA,
-//         "black" => Color::BLACK,
-//         "light_white" => Color::LIGHT_WHITE,
-//         "light_red" => Color::LIGHT_RED,
-//         "light_green" => Color::LIGHT_GREEN,
-//         "light_blue" => Color::LIGHT_BLUE,
-//         "light_yellow" => Color::LIGHT_YELLOW,
-//         "light_cyan" => Color::LIGHT_CYAN,
-//         "light_magenta" => Color::LIGHT_MAGENTA,
-//         "light_black" => Color::LIGHT_BLACK,
-//         _ => Color::default(),
-//     }
-// }
 
 /// Returns a config with values from :
 ///
