+++ src/lib.rs
@@ -2,7 +2,7 @@ pub mod action_map;
 pub mod args;
 pub mod bulkrename;
 pub mod cli_info;
-pub mod color_cache;
+pub mod colors;
 pub mod completion;
 pub mod compress;
 pub mod config;
